--- ui/litellm-dashboard/eslint-any-budget.json
@@ -1,4 +1,0 @@
-{
-  "max": 2040,
-  "target": 1500
-}
