+++ ui/litellm-dashboard/eslint-budgets.json
@@ -1,0 +1,5 @@
+{
+  "@typescript-eslint/no-explicit-any": { "max": 2040, "target": 1500 },
+  "complexity": { "max": 140, "target": 80 },
+  "max-depth": { "max": 70, "target": 30 }
+}
